Separated resolver and judger
